use apa style for citations
This style makes the in-text citations display the author name and year
of publication (instead of just numerals in square brackets). This style
makes more sense because it displays the author name by default along
with the publication year. We won't be citing a ton of things all the
time so it's OK to be a little verbose like this.