adjust code block style and formatting
Use a coloring scheme wher blue link names are pointing to the
green (child source block), and green link names are pointing to the
blue (parent) location where this child block is referenced. Change the
background color on hover to indicate where we are jumping to.

Also do not enumerate child source block names. Instead just replace
with the unicode character for a link symbol.