add link anchors for tail polyblocks
Only the leading polyblock had a proper link anchor; the other ones just
had "##deadlink" because these blocks did not have an "id" field in the
"<pre ...>" tag. This is by design because we manually inject a "#+name:
..." for only the head polyblock in `codex-UID-for-all-polyblocks`,
because when we link to child polyblocks from a parent block, we have to
choose the leader (because presumably the reader would want to start by
reading the first one).

This causes a problem for link anchors because the tail polyblocks do
not come with an HTML "id" field linke the head polyblock. We now detect
the absence of such fields and then create an id based on the source
code block name, using a hash table to keep count of how many of these
exist.

Also, delete "listing-number" because we don't use them.