problems: use a codex.theme file instead of inline HTML
This way we can just symlink to a single "codex.theme" file, regardless
of whatever CSS or JS changes we want to make across the board.

For simplicity, we don't use the codex.theme file in the toplevel
README.org file, but this OK because there's only 1 toplevel
file (wheras there will be many more problem/*/README.org files).

This is a pure refactor (no functional change).