de-anonymize some code blocks
There's no reason why these blocks should remain anonymous (such that
their link anchors get named like "#anonymous-1", "#anonymous-2", etc).

The convention is to use the same string from "#+caption: ..." in the
"#+name: ..." part. The name can differ (if desired), but the caption is
what will be displayed to the user. The "#+name: ..." is thus only
visible to Org-mode, not to the human reader browsing the HTML output.

The "#+name: ..." should be unique to each code block, because Org mode
allows for calling source code blocks by name. And so, this is at least
one reason why we refrain from treating the name and caption as
equivalent fields.