update nixpkgs channel
This resulted in some churn due to new linting errors.

Surprisingly, the CSS output also appears to be slightly different
because the lines

    .org-ns-working-text {
      /* ns-working-text-face */
      text-decoration: underline;
    }

were deleted even though we did not make any changes to the CSS
generation logic. Presumably this is because we ran this on a
NixOS (Linux) machine and not on Mac where we created some of the
previous commits.