temporarily fix for mathematcal appendix
Previously we did not encode the fact that the mathematical appendix
depends on the new "twos-complement.org" file, such that updating the
latter did not update the former. We encode this relationship with this
change.

However, this fix is not permanent because we now get this warning from
Make:

    Makefile:84: warning: overriding recipe for target 'appendix/mathematics/README.html'
    Makefile:76: warning: ignoring old recipe for target 'appendix/mathematics/README.html'

We should fix this warning in a follow-up change.