maximum subarray: handle input with only zeroes or negative numbers
Before, we would return None for inputs with only negative integers or
zeros. Instead, return the largest element found in such cases. This
only results in a small modification to Kadane's algorithm (dp()), while
exposing a flaw with dp_running_sum() which now requires special
pre-processing for this case.

We use None for the initial value of the maximum sum found so far,
instead of float("-inf") as it is used in other implementations, because
this keeps the type checker happy (otherwise it will complain about us
mixing floats with ints). This results in a small amount of extra
housekeeping work, but keeping the types clean like this is better than
dropping the arg-type check altogether.

Interestingly, dac_linear() could already gracefully handle such input,
so this change simplifies the handling of None-ness for that algorithm.